refactor(main): move desktop platform db to background/
