+++ main/src/platform/desktop/background/db/compat.ts
