+++ main/src/platform/desktop/background/db/types.ts
