+++ main/src/platform/desktop/background/platform.ts
@@ -1,5 +1,5 @@
 import type { LazyAsync } from "@/features/utils";
-import type { Database } from "../db";
+import type { Database } from "./db";
 
 export class DesktopPlatformBackground {
   constructor(public readonly db: LazyAsync<Database>) {}
